requirements.txt
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
Flask==2.0.2
Flask-PyMongo==2.3.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pymongo==4.0.1
python-dotenv==0.19.2
six==1.16.0
Werkzeug==2.0.2
